changed datatype from float64 to float32.

+++ main.go
@@ -8,15 +8,15 @@ func main() {
 	var t Table
 
 	t.Header = []string{"A", "B", "C"}
-	t.Rows = append(t.Rows, []float64{1, 2, 3})
-	t.Rows = append(t.Rows, []float64{3, 4, 5})
-	t.Rows = append(t.Rows, []float64{6, 7, 8})
-	t.Rows = append(t.Rows, []float64{17, 72, 78})
-	t.Rows = append(t.Rows, []float64{1, 2, 3})
-	t.Rows = append(t.Rows, []float64{1, 2, 3})
-	t.Rows = append(t.Rows, []float64{1, 2, 3})
-	t.Rows = append(t.Rows, []float64{100, 200, 300})
-	t.Rows = append(t.Rows, []float64{0, 0, 0})
+	t.Rows = append(t.Rows, []float32{1, 2, 3})
+	t.Rows = append(t.Rows, []float32{3, 4, 5})
+	t.Rows = append(t.Rows, []float32{6, 7, 8})
+	t.Rows = append(t.Rows, []float32{17, 72, 78})
+	t.Rows = append(t.Rows, []float32{1, 2, 3})
+	t.Rows = append(t.Rows, []float32{1, 2, 3})
+	t.Rows = append(t.Rows, []float32{1, 2, 3})
+	t.Rows = append(t.Rows, []float32{100, 200, 300})
+	t.Rows = append(t.Rows, []float32{0, 0, 0})
 
 	var model KMeansModel
 	model.Data = t
